requirement.txt
aioredis==1.3.1
amqp==2.5.2
APScheduler==3.6.3
apturl==0.5.2
asgi-redis==1.4.3
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
asttokens==2.0.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
autopep8==1.4.4
beautifulsoup4==4.8.1
billiard==3.6.2.0
bleach==2.1.2
Brlapi==0.6.6
bs4==0.0.1
certifi==2019.9.11
cffi==1.14.0
channels==2.3.1
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
daphne==2.4.0
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.8.1
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0
docutils==0.15.2
elasticsearch==7.1.0
entrypoints==0.2.3.post1
hiredis==1.0.1
html5lib==0.999999999
httpie==0.9.8
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
material==0.1
mccabe==0.6.1
mistune==0.8.3
MouseInfo==0.1.2
msgpack==0.6.2
msgpack-python==0.5.6
mypy==0.740
mypy-extensions==0.4.3
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.17.0
oauth==1.0.1
olefile==0.45.1
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.0
playsound==1.2.2
prompt-toolkit==1.0.15
protobuf==3.0.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyGetWindow==0.0.7
Pygments==2.2.0
pygobject==3.26.1
PyHamcrest==2.0.0
pylint==2.4.4
pymacaroons==0.13.0
pymongo==3.8.0
PyMsgBox==1.0.7
pymssql==2.1.4
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyperclip==1.7.0
PyRect==0.1.4
pyRFC3339==1.0
PyScreeze==0.1.24
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
redis==3.3.11
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
schedule==0.6.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.6
sqlparse==0.3.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.7
testpath==0.3.1
thonny==3.2.3
tornado==4.5.3
traitlets==4.3.2
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.0
typing-extensions==3.7.4.1
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
vine==1.3.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
zipp==3.0.0
zope.interface==4.7.1
